feat: develop get user info endpoint

+++ Kazka.Application/Features/Users/Commands/Update/UpdateUserCommand.cs
@@ -6,7 +6,7 @@ namespace Kazka.Application.Features.User.Command.Update
     public record UpdateUserCommand
     (
         string? Name,
-        uint? Age,
+        byte? Age,
         string? ProfilePictureUrl
     ): IRequest<UserResponse>;
 }
